requirements.txt
# Mid-term Stock Planner Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
lightgbm>=4.0.0
scikit-learn>=1.3.0

# Explainability
shap>=0.42.0

# Configuration
pyyaml>=6.0

# Statistics and scientific computing
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0

# Optional: Technical analysis library (faster indicator computation)
# ta>=0.10.0  # Uncomment if you want to use the ta library

# Optional: Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Jupyter notebook support
# jupyter>=1.0.0
# ipywidgets>=8.0.0